requirements.txt
discord.py==2.3.2
redis==4.5.4
pytz==2022.6
python-dotenv==0.21.0
